requirements.txt
python-dotenv
google-adk
litellm
requests
beautifulsoup4
tiktoken
